updated variable num from int to Long.

+++ Client.h
@@ -15,6 +15,6 @@ public:
 	}*/
 
 	string firstname, lastName;
-	int mobile;
+	long mobile;
 };
